added inventory and owner pages

+++ src/sql/remove_procedures.sql
@@ -54,8 +54,7 @@ CREATE PROCEDURE remove_ingredient(
 )
 BEGIN
 
-    DECLARE ing_id;
-
+    DECLARE ing_id INT;
 
     SELECT ingredient_id
     INTO ing_id
@@ -63,12 +62,67 @@ BEGIN
     WHERE label = p_label
     LIMIT 1;
 
-    DELETE FROM menuitems
-    JOIN menuitemingredients
-    ON menuitems.item_id = menuitemingredients.menu_item_id
-    WHERE menuitemingredients.ingredientIid = ing_id;
+    IF ing_id IS NOT NULL THEN
+        DELETE FROM menuitemingredients
+        WHERE ingredient_id = ing_id;
+
+        -- remove from ingredients table
+        DELETE FROM ingredients
+        WHERE ingredient_id = ing_id;
+    END IF;
+END$$
 
 
+CREATE PROCEDURE remove_menu_item(
+    IN p_title VARCHAR(255)
+)
+BEGIN
+
+    DECLARE it_id INT;
+
+    SELECT item_id
+    INTO it_id
+    FROM menuitems
+    WHERE title = p_title
+    LIMIT 1;
+
+    IF it_id IS NOT NULL THEN
+
+        -- remove from menuitemingredients
+        DELETE FROM menuitemingredients
+        WHERE menu_item_id = it_id;
+
+        -- remove from menu items
+        DELETE FROM menuitems
+        WHERE item_id = it_id;
+    END IF; 
+
+END$$
+
+
+CREATE PROCEDURE remove_customer(
+    IN p_phone VARCHAR(255)
+)
+BEGIN
+    DECLARE c_id INT;
+
+    -- get customer id
+    SELECT customer_id 
+    INTO c_id
+    FROM customers
+    WHERE phone_num = p_phone
+    LIMIT 1;
+
+    IF c_id IS NOT NULL THEN
+        -- edit affected sales
+        UPDATE sales
+        SET customer_id = NULL
+        WHERE customer_id = c_id;
+
+        -- remove from customers table
+        DELETE FROM customers
+        WHERE customer_id = c_id;
+    END IF;
 END$$
 
 DELIMITER ;
